[finishes #105667918] adding more accurate error handling for ops manager connection
when connecting to the ops manager api,
we were not checking the statuscode in the
response and taking error action.

we were leaning on the parsing of the response to
throw an error. this was leading to missleading/vague
errors. we are now returning the error message
from ops manager when the status code is anything
but statusOK.