[#108624580] patching v1.6 postgres multi-vendoring bug
in 1.6 there seem to be multiple versions of postgres utils
the version we had been using, while still there, does not
work to backup/restore the version of postgres which is
running.

in the event of a 1.6 flag set in the ENV VAR
then we will use the new version of the postgres utils which map
to psql 9.4.2.