[#112077185]  refactoring of tests and clarity of intent in iaasparser
er builder tests were swallowing errors and testing incorrect
values, we refactored this to test the proper error output

refactored iaas object methods to return a more clear
hasSSHKey value. the prvious error value returned did not
clearly identify the intent of the response and the state
status it was returning.